Introduccion de datos en la bd

+++ app/src/main/java/es/udc/psi14/prueba/Sensor.java
@@ -4,21 +4,21 @@ package es.udc.psi14.prueba;
  * Created by Santiago on 26/11/2014.
  */
 public class Sensor {
-    private int medida;
+    private float medida;
     private String nombre,identificador;
     private long id, fecha;
 
 
     //TODO: añandir coordenadas gps?
 
-    public Sensor(int medida, String nombre, String identificador, long fecha ){
+    public Sensor(float medida, String nombre, String identificador, long fecha ){
         this.medida=medida;
         this.nombre=nombre;
         this.identificador=identificador;
         this.fecha=fecha;
     }
 
-    public Sensor(int medida, String nombre, String identificador, long fecha, long id ){
+    public Sensor(float medida, String nombre, String identificador, long fecha, long id ){
         this.medida=medida;
         this.nombre=nombre;
         this.identificador=identificador;
@@ -46,7 +46,7 @@ public class Sensor {
         this.id=id;
     }
 
-    public int getMedida(){
+    public float getMedida(){
         return medida;
     }
